requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyMySQL==1.1.0

langchain==0.3.7
langchain-core==0.3.15
langchain-google-genai==2.0.0
google-generativeai==0.7.2

SQLAlchemy==2.0.23
